Refactor: 프로필 페이지 컴포넌트의 프로필 로직을 ProfileCard로 분리하여 관리하기 쉽도록 변경 Profile 컴포넌트에서 렌더링되는 컴포넌트가 여러곳을 참조하기 때문에 관리 및 가독성을 위하여 유저 Profile의 로직도 ProfileCard로 분리함 Resolves: #119
