+++ src/routes/Router.jsx
@@ -41,11 +41,11 @@ export default function Router() {
 					<Route path='/profile/' element={<Outlet />}>
 						<Route path='' element={<Profile />} />
 						<Route path='*' element={<Page404 />} />
+						<Route path='edit/' element={<MyProfileEdit />} />
 						<Route path=':accountUsername' element={<Outlet />}>
 							<Route path='' element={<Profile />} />
 							<Route path='*' element={<Page404 />} />
 							<Route path=':follow/' element={<Follow />}></Route>
-							<Route path='edit/' element={<MyProfileEdit />} />
 						</Route>
 					</Route>
